Refactor PowerShell runspace initialization logic for improved modularity and reuse
