+++ PoshMcp.Tests/Integration/CommandLineTests.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace PoshMcp.Tests
+namespace PoshMcp.Tests.Integration
 {
     /// <summary>
     /// Tests for the command line argument parsing functionality
@@ -22,24 +22,24 @@ namespace PoshMcp.Tests
         {
             // This test verifies that the help functionality works
             // We can't easily test the exact output, but we can ensure the command doesn't crash
-            
+
             // Since Program.Main returns int now, we'll test that the command line structure is valid
             // by checking that the help text generation doesn't throw exceptions
-            
+
             var rootCommand = new RootCommand("PowerShell MCP Server - Provides access to PowerShell commands via Model Context Protocol");
-            
+
             var evaluateToolsOption = new Option<bool>(
                 aliases: new[] { "--evaluate-tools", "-e" },
                 description: "Evaluate and list discovered PowerShell tools without starting the MCP server");
-            
+
             var verboseOption = new Option<bool>(
                 aliases: new[] { "--verbose", "-v" },
                 description: "Enable verbose logging");
-                
+
             var debugOption = new Option<bool>(
                 aliases: new[] { "--debug", "-d" },
                 description: "Enable debug logging");
-                
+
             var traceOption = new Option<bool>(
                 aliases: new[] { "--trace", "-t" },
                 description: "Enable trace logging");
@@ -52,7 +52,7 @@ namespace PoshMcp.Tests
             // If we get here without exceptions, the command structure is valid
             Assert.NotNull(rootCommand);
             Assert.Equal(4, rootCommand.Options.Count);
-            
+
             // Verify option names
             Assert.Contains(rootCommand.Options, o => o.HasAlias("--evaluate-tools"));
             Assert.Contains(rootCommand.Options, o => o.HasAlias("--verbose"));
@@ -73,19 +73,19 @@ namespace PoshMcp.Tests
         {
             // Test that all our command line options are properly recognized
             var rootCommand = new RootCommand("PowerShell MCP Server - Provides access to PowerShell commands via Model Context Protocol");
-            
+
             var evaluateToolsOption = new Option<bool>(
                 aliases: new[] { "--evaluate-tools", "-e" },
                 description: "Evaluate and list discovered PowerShell tools without starting the MCP server");
-            
+
             var verboseOption = new Option<bool>(
                 aliases: new[] { "--verbose", "-v" },
                 description: "Enable verbose logging");
-                
+
             var debugOption = new Option<bool>(
                 aliases: new[] { "--debug", "-d" },
                 description: "Enable debug logging");
-                
+
             var traceOption = new Option<bool>(
                 aliases: new[] { "--trace", "-t" },
                 description: "Enable trace logging");
@@ -105,7 +105,7 @@ namespace PoshMcp.Tests
                     break;
                 }
             }
-            
+
             Assert.True(optionExists, $"Option '{option}' should be recognized");
         }
 
@@ -114,11 +114,11 @@ namespace PoshMcp.Tests
         {
             // Test that we can combine options like --evaluate-tools --verbose
             var rootCommand = new RootCommand("PowerShell MCP Server - Provides access to PowerShell commands via Model Context Protocol");
-            
+
             var evaluateToolsOption = new Option<bool>(
                 aliases: new[] { "--evaluate-tools", "-e" },
                 description: "Evaluate and list discovered PowerShell tools without starting the MCP server");
-            
+
             var verboseOption = new Option<bool>(
                 aliases: new[] { "--verbose", "-v" },
                 description: "Enable verbose logging");
@@ -128,10 +128,10 @@ namespace PoshMcp.Tests
 
             // Test parsing combined options
             var parseResult = rootCommand.Parse(new[] { "--evaluate-tools", "--verbose" });
-            
+
             Assert.NotNull(parseResult);
             Assert.Empty(parseResult.Errors);
-            
+
             // Verify both options are recognized
             Assert.True(parseResult.GetValueForOption(evaluateToolsOption));
             Assert.True(parseResult.GetValueForOption(verboseOption));
